add dynamic pagination in member jsp

+++ src/main/java/com/mypractice/assistancetracker/service/MemberService.java
@@ -34,6 +34,6 @@ public interface MemberService {
 	/**
 	 * @return
 	 */
-	Long getMemeberPageCount();
+	int getMemeberPageCount();
 
 }
